Fix the folder to create the configurations for the different environments
